Reduce Duplication of Lines Code
